pyproject.toml
[project]
name = "sqlhunt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amqp==5.3.1",
    "asgiref==3.8.1",
    "billiard==4.2.1",
    "celery==5.5.3",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "cryptography==44.0.3",
    "defusedxml==0.7.1",
    "dj-rest-auth==4.0.0",
    "django>=5.2.2",
    "django-allauth==0.52.0",
    "django-cors-headers==4.7.0",
    "django-defender>=0.9.8",
    "django-extensions==4.1",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.5.0",
    "faker==37.3.0",
    "gunicorn>=23.0.0",
    "idna==3.10",
    "kombu==5.5.4",
    "markupsafe==3.0.2",
    "oauthlib==3.2.2",
    "packaging==25.0",
    "prompt-toolkit==3.0.51",
    "psycopg2-binary==2.9.10",
    "pycparser==2.22",
    "pyjwt==2.9.0",
    "pyopenssl==25.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python3-openid==3.2.0",
    "redis>=6.4.0",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "six==1.17.0",
    "sqlalchemy==2.0.41",
    "sqlglot==26.25.3",
    "sqlparse==0.5.3",
    "tblib>=3.1.0",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "werkzeug==3.1.3",
]
